# backend/api-docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing scheduled tasks and task groups",
        "title": "Cron Observer API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/projects/{project_id}/tasks": {
            "post": {
                "description": "Create a new scheduled task in a project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Create a new task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Task creation request",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Task"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/projects/{project_id}/tasks/{task_uuid}": {
            "put": {
                "description": "Update an existing scheduled task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Update a task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task UUID",
                        "name": "task_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Task update request",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Task"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an existing scheduled task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Delete a task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task UUID",
                        "name": "task_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.CreateTaskRequest": {
            "type": "object",
            "required": [
                "name",
                "project_id",
                "schedule_config",
                "schedule_type",
                "trigger_config"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 1000
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "project_id": {
                    "type": "string"
                },
                "schedule_config": {
                    "$ref": "#/definitions/models.ScheduleConfig"
                },
                "schedule_type": {
                    "enum": [
                        "RECURRING",
                        "ONEOFF"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ScheduleType"
                        }
                    ]
                },
                "status": {
                    "enum": [
                        "ACTIVE",
                        "PAUSED",
                        "DISABLED"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TaskStatus"
                        }
                    ]
                },
                "task_group_id": {
                    "description": "Optional task group ID",
                    "type": "string"
                },
                "trigger_config": {
                    "$ref": "#/definitions/models.TriggerConfig"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "project_id is required"
                    ]
                },
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                }
            }
        },
        "models.Frequency": {
            "type": "object",
            "required": [
                "unit",
                "value"
            ],
            "properties": {
                "unit": {
                    "description": "Unit: \"s\" (seconds), \"m\" (minutes), \"h\" (hours)",
                    "enum": [
                        "s",
                        "m",
                        "h"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.FrequencyUnit"
                        }
                    ]
                },
                "value": {
                    "description": "Numeric value (e.g., 15)",
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "models.FrequencyUnit": {
            "type": "string",
            "enum": [
                "s",
                "m",
                "h"
            ],
            "x-enum-varnames": [
                "FrequencyUnitSecond",
                "FrequencyUnitMinute",
                "FrequencyUnitHour"
            ]
        },
        "models.HTTPTriggerConfig": {
            "type": "object",
            "required": [
                "method",
                "url"
            ],
            "properties": {
                "body": {},
                "headers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "method": {
                    "type": "string"
                },
                "timeout": {
                    "type": "integer",
                    "maximum": 300,
                    "minimum": 1
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "models.ScheduleConfig": {
            "type": "object",
            "required": [
                "timezone"
            ],
            "properties": {
                "cron_expression": {
                    "description": "If provided, TimeRange and DaysOfWeek are ignored",
                    "type": "string"
                },
                "days_of_week": {
                    "description": "Used only if CronExpression is not provided",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "exclusions": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "time_range": {
                    "description": "Used only if CronExpression is not provided",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TimeRange"
                        }
                    ]
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
        "models.ScheduleType": {
            "type": "string",
            "enum": [
                "RECURRING",
                "ONEOFF"
            ],
            "x-enum-varnames": [
                "ScheduleTypeRecurring",
                "ScheduleTypeOneOff"
            ]
        },
        "models.Task": {
            "description": "Task represents a scheduled task entity",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2025-01-15T10:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Backup database daily"
                },
                "id": {
                    "type": "string",
                    "example": "507f1f77bcf86cd799439011"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "name": {
                    "type": "string",
                    "example": "Daily Backup"
                },
                "project_id": {
                    "type": "string",
                    "example": "507f1f77bcf86cd799439011"
                },
                "schedule_config": {
                    "$ref": "#/definitions/models.ScheduleConfig"
                },
                "schedule_type": {
                    "enum": [
                        "RECURRING",
                        "ONEOFF"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ScheduleType"
                        }
                    ],
                    "example": "RECURRING"
                },
                "status": {
                    "enum": [
                        "ACTIVE",
                        "PAUSED",
                        "DISABLED"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TaskStatus"
                        }
                    ],
                    "example": "ACTIVE"
                },
                "task_group_id": {
                    "description": "Optional reference to task group",
                    "type": "string",
                    "example": "507f1f77bcf86cd799439011"
                },
                "trigger_config": {
                    "$ref": "#/definitions/models.TriggerConfig"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-01-15T10:00:00Z"
                },
                "uuid": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "models.TaskStatus": {
            "type": "string",
            "enum": [
                "ACTIVE",
                "PAUSED",
                "DISABLED"
            ],
            "x-enum-varnames": [
                "TaskStatusActive",
                "TaskStatusPaused",
                "TaskStatusDisabled"
            ]
        },
        "models.TimeRange": {
            "type": "object",
            "required": [
                "end",
                "frequency",
                "start"
            ],
            "properties": {
                "end": {
                    "description": "Format: \"HH:MM\"",
                    "type": "string"
                },
                "frequency": {
                    "description": "Frequency with value and unit (e.g., {value: 15, unit: \"m\"})",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Frequency"
                        }
                    ]
                },
                "start": {
                    "description": "Format: \"HH:MM\"",
                    "type": "string"
                }
            }
        },
        "models.TriggerConfig": {
            "type": "object",
            "required": [
                "http",
                "type"
            ],
            "properties": {
                "http": {
                    "$ref": "#/definitions/models.HTTPTriggerConfig"
                },
                "type": {
                    "enum": [
                        "HTTP"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TriggerType"
                        }
                    ]
                }
            }
        },
        "models.TriggerType": {
            "type": "string",
            "enum": [
                "HTTP"
            ],
            "x-enum-varnames": [
                "TriggerTypeHTTP"
            ]
        },
...
        "models.UpdateTaskRequest": {
            "type": "object",
            "required": [
                "name",
                "schedule_config",
                "schedule_type",
                "trigger_config"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 1000
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "schedule_config": {
                    "$ref": "#/definitions/models.ScheduleConfig"
                },
                "schedule_type": {
                    "enum": [
                        "RECURRING",
                        "ONEOFF"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ScheduleType"
                        }
                    ]
                },
                "status": {
                    "enum": [
                        "ACTIVE",
                        "PAUSED",
                        "DISABLED"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TaskStatus"
                        }
                    ]
                },
                "task_group_id": {
                    "description": "Optional task group ID",
                    "type": "string"
                },
                "trigger_config": {
                    "$ref": "#/definitions/models.TriggerConfig"
                }
            }
        }
    }
}